refactor: reorganize type imports and enhance type definitions across components
