+++ src/views/sidebar/commons/folders-accordion.tsx
@@ -13,10 +13,9 @@ import {
 	AccordionSummary as MUIAccordionSummary,
 	AccordionDetails as MUIAccordionDetails
 } from '@mui/material';
-import { FOLDERS, hasId, theme } from '@zextras/carbonio-ui-commons';
+import { Folder, FOLDERS, hasId, theme } from '@zextras/carbonio-ui-commons';
 
 import { isRoot, isSpam, isTrash } from 'helpers/folders';
-import { Folder } from 'types/index.d';
 
 type FolderAccordionProps = {
 	folders: Array<Folder>;
